fix(richtext-lexical): editor placeholder had incorrect padding set for small viewports
