+++ packages/richtext-lexical/src/lexical/LexicalEditor.tsx
@@ -23,9 +23,10 @@ import { LexicalContentEditable } from './ui/ContentEditable.js'
 export const LexicalEditor: React.FC<
   {
     editorContainerRef: React.RefObject<HTMLDivElement | null>
+    isSmallWidthViewport: boolean
   } & Pick<LexicalProviderProps, 'editorConfig' | 'onChange'>
 > = (props) => {
-  const { editorConfig, editorContainerRef, onChange } = props
+  const { editorConfig, editorContainerRef, isSmallWidthViewport, onChange } = props
   const editorConfigContext = useEditorConfigContext()
   const [editor] = useLexicalComposerContext()
 
@@ -77,24 +78,6 @@ export const LexicalEditor: React.FC<
       editorConfigContext.parentEditor?.unregisterChild?.(editorConfigContext.uuid)
     }
   }, [editor, editorConfigContext])
-
-  const [isSmallWidthViewport, setIsSmallWidthViewport] = useState<boolean>(false)
-
-  useEffect(() => {
-    const updateViewPortWidth = () => {
-      const isNextSmallWidthViewport = window.matchMedia('(max-width: 768px)').matches
-
-      if (isNextSmallWidthViewport !== isSmallWidthViewport) {
-        setIsSmallWidthViewport(isNextSmallWidthViewport)
-      }
-    }
-    updateViewPortWidth()
-    window.addEventListener('resize', updateViewPortWidth)
-
-    return () => {
-      window.removeEventListener('resize', updateViewPortWidth)
-    }
-  }, [isSmallWidthViewport])
 
   return (
     <React.Fragment>
